mfu_agent/pyproject.toml
[project]
name = "mfu-agent"
version = "0.1.0"
description = "AI-agent for MFU fleet health monitoring with integral health index calculation"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.38,<2",
    "pydantic>=2.0,<3",
    "qdrant-client>=1.11,<2",
    "FlagEmbedding>=1.3,<2",
    "sentence-transformers>=3.0,<4",
    "openai>=1.40,<2",
    "PyYAML>=6.0,<7",
    "jinja2>=3.1,<4",
    "weasyprint>=62,<63",
    "pymupdf>=1.24,<2",
    "pandas>=2.2,<3",
    "tiktoken>=0.7,<1",
    "python-dateutil>=2.9,<3",
    "chardet>=5.0,<6",
    "openpyxl>=3.1,<4",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5.0,<6",
    "ruff>=0.6,<1",
    "mypy>=1.11,<2",
    "Pillow>=10.0,<12",
]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agent*", "tools*", "rag*", "data_io*", "reporting*", "llm*", "config*", "state*", "pages*", "ui*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = ["E501", "RUF001", "RUF002", "RUF003"]

[tool.ruff.lint.isort]
known-first-party = ["agent", "tools", "rag", "data_io", "reporting", "llm", "config", "state"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "streamlit.*",
    "sentence_transformers.*",
    "FlagEmbedding.*",
    "qdrant_client.*",
    "weasyprint.*",
    "fitz.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests", "tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false
disallow_untyped_decorators = false
warn_return_any = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "e2e: end-to-end tests (full pipeline with mock LLM)",
]
